% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnl_fd_ova.R
\name{mnl_fd_ova}
\alias{mnl_fd_ova}
\title{Multinomial First Differences Prediction (Observed Value Approach)}
\usage{
mnl_fd_ova(model, data, xvari, scenname, scenvalues, by = 1,
  nsim = 1000, seed = "random", probs = c(0.025, 0.975))
}
\description{
Multinomial First Differences Prediction (Observed Value Approach)
}
\examples{

}
