% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnlpredOVA.R
\name{mnlpredOVA}
\alias{mnlpredOVA}
\title{Title}
\usage{
mnlpredOVA(model, data, xvari, scenname = NULL, scenvalue = NULL,
  by = 1, nsim = 1000, seed = "random", probs = c(0.025, 0.975))
}
\arguments{
\item{probs}{}
}
\description{
Title
}
